\name{diagnose-methods}
\docType{methods}
\alias{diagnose-methods}
\alias{diagnose,ppm-method}
\alias{diagnose,ppmlasso-method}
\title{Methods for function \code{diagnose}}
\description{
Methods for function \code{\link{diagnose}}
}
\section{Methods}{
\describe{

\item{\code{signature(fit = "ppm")}}{
Creates residual plots for a \code{ppm} object. See the help function for \code{diagnose.ppm} in \code{spatstat} for more details.
}

\item{\code{signature(fit = "ppmlasso")}}{
Creates residual plots for a \code{ppmlasso} object. See the help function for \code{diagnose.ppm} in \code{spatstat} for more details.
}
}}
\keyword{methods}
