\name{envelope-methods}
\docType{methods}
\alias{envelope-methods}
\alias{envelope,ppmlasso-method}
\title{Methods for function \code{envelope}}
\description{
Methods for function \code{\link{envelope}}
}
\section{Methods}{
\describe{

\item{\code{signature(Y = "ppmlasso")}}{
Creates Monte Carlo simulation envelopes for a given function for a \code{ppmlasso} object. See the help function for \code{envelope} in \code{spatstat} for more details.
}
}}
\keyword{methods}
