\name{predict-methods}
\docType{methods}
\alias{predict-methods}
\alias{predict,ppmlasso-method}
\title{Methods for function \code{predict}}
\description{
Methods for function \code{\link{predict}}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ppmlasso")}}{
Creates a vector of predicted intensities for an object of class \code{ppmlasso}.
}
}}
\keyword{methods}
