\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,ppmlasso-method}
\title{Methods for function \code{print}}
\description{
Methods for function \code{\link{print}}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ppmlasso")}}{
Prints output for a \code{ppmlasso} object with details controlled by arguments of the 
\code{\link{print.ppmlasso}} function.
}
}}
\keyword{methods}
